+++ src/adminsite/fields/__init__.py
@@ -2,6 +2,7 @@ from adminsite.fields.base import Field
 from adminsite.fields.choice import ChoiceField
 from adminsite.fields.computed import Computed
 from adminsite.fields.files import FileField, ImageField
+from adminsite.fields.json_field import JSONField
 from adminsite.fields.registry import (
     FieldRegistry,
     build_default_registry,
@@ -34,6 +35,7 @@ __all__ = [
     "FloatField",
     "ImageField",
     "IntegerField",
+    "JSONField",
     "RelationField",
     "StringField",
     "TextField",